first-cdk-deployment/requirements.txt
aws-cdk-lib==2.178.1
constructs>=10.0.0,<11.0.0
